re #2452, Renamed users.py to user.py to be consitent with other view file names.
